+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    See: https://github.com/elBukkit/MagicWorlds/tree/master/src/main/resources/examples/mixins for some fun examples!
  - Add per-world resource pack feature
  - Add WGCustomFlags integration, "spawn-tags" custom flag and tags on spawn rules
+ - WorldGuard custom flag support added, requires latest builds of WorldGuard, replaces WGCustomFlags
  - The default config is now empty! See the "mixins" folder for the chest populators
 
 # 1.3